requirements.txt
opencv-python==4.8.0.74
pyserial==3.5